Use last selected root id for the global inserter in widgets editor
